src/plagarism-detector/requirements.txt
torch
transformers
scikit-learn
numpy
tqdm
boto3
python-dotenv